Accept emails bound for myorigin if set
If myorigin is set to something other than the default '$myhostname', we
need to also accept that in mydestination. If not, mail to local
recipients like 'root' will get sent via relayhost to root@<myorigin>
and will not hit the alias in /etc/aliases

NOTE: This has the unfortunate side effect of duplicating hostname if
myorigin is not set, e.g. if

- facter $hostname resolves to mon2021,
- facter $fqdn resolves to mon2021.test.rc.nectar.org.au
then
- postfix $myhostname will be mon2021.test.rc.nectar.org.au [1]

and the config will end up looking like

mydestination = $myhostname, mon2021, mon2021.test.rc.nectar.org.au, ...

[1] https://www.postfix.org/postconf.5.html#myhostname

Change-Id: I63cc3ba821ef41addce08453ef9cff9c2ff2fe3f