Add firewall support for postfix
Add nectar firewall support to the postfix module, so we can automatically
add the necessary rules for allow incoming SMTP for any external networks
provided.

We specifically move the "localhost" CIDRs to become non-optional,
so we can more easily split them out to we're only creating rules
for the external networks we need to.

Change-Id: Ic91023b772df1114e5788ec67c48dc3aeb9f4e4a